# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AgentMux Public Integration API
  description: >
    Public integration contract for AgentMux. Stability tiers (stable / beta /
    internal) are defined in CONTRACT.md; only stable and beta endpoints are
    described here. The OpenAI-compatible surface (/v1/responses, /v1/files)
    follows the upstream OpenAI Responses specification and is intentionally
    not duplicated in this document.
  version: "1.3"
  x-contract-version: "1.3"
servers:
  - url: http://127.0.0.1:8765
    description: Local AgentMux daemon (default listen address)
security:
  - bridgeToken: []
...
paths:
...
  /api/v1/orchestrations:
    get:
      tags: [orchestrations]
      operationId: getOrchestrations
      summary: List orchestrations or fetch one by id
      parameters:
        - name: id
          in: query
          required: false
          schema: { type: string }
        - name: active
          in: query
          required: false
          schema: { type: boolean }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Orchestration or orchestration list
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Orchestration"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Orchestration"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [orchestrations]
      operationId: createOrchestration
      summary: Create and start a multi-agent DAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tasks]
              properties:
                name:
                  type: string
                max_concurrency:
                  type: integer
                  description: 1..12, default 4
                tasks:
                  type: array
                  maxItems: 32
                  items:
                    $ref: "#/components/schemas/OrchestrationTask"
      responses:
        "202":
          description: Accepted orchestration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orchestration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bridgeToken:
      type: http
      scheme: bearer
      description: >
        The [bridge] token from config.toml. Required on /api/* and /v1/*
        whenever bridge.enabled = true. A valid agentmux_console session
        cookie is an equivalent credential for browser-embedded Console use.
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid bridge token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Orchestration:
      type: object
      required: [id, status, max_concurrency, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        status:
          $ref: "#/components/schemas/OrchestrationStatus"
        max_concurrency: { type: integer }
        error: { type: string }
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/OrchestrationTask"
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrchestrationTask:
      type: object
      required: [id, input]
      properties:
        id: { type: string }
        orchestration_id: { type: string }
        agent_id: { type: string }
        project: { type: string }
        input: { type: string }
        depends_on:
          type: array
          items: { type: string }
        status:
          $ref: "#/components/schemas/OrchestrationStatus"
        output: { type: string }
        error: { type: string }
        invocation_id: { type: string }
        conversation_id: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrchestrationStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
...